fixed versioning bug for new users

+++ src/routes/(app)/+layout.svelte
@@ -45,7 +45,8 @@
 	const getDB = async () => {
 		const DB = await openDB("Chats", 2, {
 			upgrade(db) {
-				db.deleteObjectStore("chats");
+				if (db.objectStoreNames.contains("chats"))
+					db.deleteObjectStore("chats");
 				const store = db.createObjectStore("chats", {
 					keyPath: "id",
 					autoIncrement: true
